Fixed refresh issue on tasklist
