+++ src/components/override-sdk/infra/Assignment/Assignment.tsx
@@ -44,7 +44,7 @@ export default function Assignment(props) {
   const { t } = useTranslation();
   const serviceShuttered = useServiceShuttered();
   const { setAssignmentPConnect }: any = useContext(StoreContext);
-  const { appBacklinkProps } = useContext(AppContext)
+  const { appBacklinkProps } = useContext(AppContext);
 
   const AssignmentCard = SdkComponentMap.getLocalComponentMap()['AssignmentCard']
     ? SdkComponentMap.getLocalComponentMap()['AssignmentCard']
@@ -98,6 +98,25 @@ export default function Assignment(props) {
       clearTimeout(updateErrorTimeOut);
     };
   }, [errorMessages]);
+
+  useEffect(() => {
+    const handleBeforeUnload = () => {
+      // Perform actions before the component unloads
+      sessionStorage.setItem('isAutocompleteRendered', 'false');
+
+      const assignmentID = thePConn.getCaseInfo().getAssignmentID();
+      sessionStorage.setItem('assignmentID', assignmentID);
+
+      PCore.getContainerUtils().closeContainerItem(
+        PCore.getContainerUtils().getActiveContainerItemContext('app/primary'),
+        { skipDirtyCheck: true }
+      );
+    };
+    window.addEventListener('beforeunload', handleBeforeUnload);
+    return () => {
+      window.removeEventListener('beforeunload', handleBeforeUnload);
+    };
+  }, []);
 
   let containerName;
 
@@ -428,19 +447,23 @@ export default function Assignment(props) {
                 attributes={{ type: 'link' }}
               ></Button>
             ) : null
-          )} 
+          )}
           {
             // If there is no previous action button, and a 'appcontext' backlink action is set, show a backlink that performs the appcontext backlink action
-            arSecondaryButtons?.findIndex(button=>button.name === 'Previous') === -1
-            && appBacklinkProps.appBacklinkAction &&  <Button
-              variant='backlink'
-              onClick={e => {
-                e.target.blur();
-                appBacklinkProps.appBacklinkAction();
-              }}
-              key='createstagebacklink'
-              attributes={{ type: 'link' }}
-            >{appBacklinkProps.appBacklinkText}</Button>       
+            arSecondaryButtons?.findIndex(button => button.name === 'Previous') === -1 &&
+              appBacklinkProps.appBacklinkAction && (
+                <Button
+                  variant='backlink'
+                  onClick={e => {
+                    e.target.blur();
+                    appBacklinkProps.appBacklinkAction();
+                  }}
+                  key='createstagebacklink'
+                  attributes={{ type: 'link' }}
+                >
+                  {appBacklinkProps.appBacklinkText}
+                </Button>
+              )
           }
           <MainWrapper>
             {errorSummary && errorMessages.length > 0 && (
